chore: removed topology suite support from project and used simple float values for points

+++ src/RideSharing.Application/TripRequests/Commands/TripRequest/TripRequestCommandDto.cs
@@ -6,8 +6,8 @@ namespace RideSharing.Application.TripRequests.Commands.TripRequests
 {
 	public record struct TripRequestCommandDto(
 		string CustomerId,
-		Tuple<double, double> Source,
-		Tuple<double, double> Destination,
+		Tuple<float, float> Source,
+		Tuple<float, float> Destination,
 		CabType CabType,
 		PaymentMethod PaymentMethod)
 		: IRequest<Result<string>>
